Get a single instrument from the instruments resource.

Get a single instrument from the instruments resource.

Usage

get_instrument(
  instruments_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)

Arguments

instruments_id

An integer.

as_data_frame

A logical for toggling whether to return results as data frame or list defaults to TRUE.

dry_run

A logical for toggling a dry run of the request, defaults to FALSE.

rate_limit

A logical for toggling automatic rate limiting based on rate limit headers, defaults to FALSE.

api_key

A valid OpenAQ API key string, defaults to NULL.

Value

A data frame or a list of the results.

Examples

if (FALSE) { # \dontrun{
instrument <- get_instrument(42)
} # }